Reimplement parser to parse into list of commands.
Previously we separated parsing into a list of commands
into two steps: parse into a list of code, then convert
this to a list of commands. In the previous commit,
we implemented this, and discussed its ambiguous nature.

After consulting the LinuxCNC G-code standard, we
found that the standard recommended that each
command be in a single line. We know assume this
is the case for the input. Before we did not care
about any whitespace.

This implementation ensures any non-standard structure
in the list of commands is due solely to the input
string having non-standard structure.

As for the implementation itself, we again parse
by using only the lexer part of parser-tools. In
particular, we use a combination of 3 lexers.
One parses a whole string of G-code into a list
of commands. The second parses a G-code line into
a command. The third parses a G-code word into a code.

This might seem tricky at first, but it is a "combinatorial"
way of doing it.