Remove other functions.
We decided to do this since they were rather useless functions
with regards to a library. The equality functions were
technically already implemented with equal?. The other
functions were simply useless in the general case.

It is more likely, the user needs functions with
specific behavior. It is best for them to implement it
themselves, and for us to provide general functions when
possible.