Add initial support for comments.
By support, we mean that comments are ignored during parsing.
Only inactive parantheses comments are supported for now.
Semicolon comments are not supported, but should be in the future.

Active comments are not supported and will likely not be supported.
Active comments are useful for human G-coders. Programs that
generate G-code have no real use for them, and often do not
emit them. Since this library is geared toward manipulating G-code
produced by programs, we will likely not supported active comments.