Fix bug with comment parsing.
There was an issue regarding the regular expression matching a parantheses
comment. For example:

(comment text)
G21 (more comment text)

should produce a command list with just a G21 command, but instead the parser
produces the empty list.

This happened because the regular expression for the comment was
(.*) where .* will accept absolute any character, including newlines!
In the case above, the newline and G21 became part of the match for the
comment. So they were ignored during parsing.