Replace name/parameter updaters with more general function.
Before we had function that would take a single command
and update the name or parameter list. These were made
primarily for use with the higher-level mapping functions
that do the same on a list of commands. This is a bit
unnecessary. It is more general to just have a high-level
mapping function that works on the entire command
rather than just its name or parameter list.