Fix bug in update-coordinates.
It is possible that one of the codes in the updated
coordinates is #f. In order to replace coordinate codes
in the original command, we make use of parameter-by-letter.
In particular, the new coordinate codes are sent to it
as an argument. However, parameter-by-letter assumes
arguments are all code?. Therefore, when #f is sent,
we get an exception.

The fix is to simply filter the list of updated coordinate
codes for non-#f values.

We could use #f to denote a removed coordinate code,
but this makes implementation more difficult and messy.
Other functions can be used to remove specific codes.