Add documentation for structs.
Two commit ago, we were supposed to say that a documentation
effort has started, but this was neglected. That being said,
a documentation effort has started! As we add more scribbles
we will remove documentation that was placed informally
within the source code files.

Also some trivial changes regarding the titles and formatting
was done in this commit.