Harden contract for code.
Before the letter needed to be just a symbol. Now
it must be a symbol that is actually used in the
G-code standard.

The current contract for code is a good balance between
rigor and flexibility. It ensures the G-code is in
a canonical format, but allows the user to
create their own codes like G1000 even if they are
not defined by the standard.