Add tests for coord? and other tests.
The other tests added were for the other coordinate checking functions.

Currently not all the tests for coord? pass. This is due a contract violation
appearing from the expression (coord? (vector 1 2 3)).

coord? checks if its argument is a vector and then passes each element to the
specified function. This is one of the other coordinate checking functions (for
example x-coord?). Those functions pass the element onto code checking functions
(for example x-code?).

The code checking function expects a code. In this case
it is given a number. The easiest fix is to change the contract
on code checking functions to expect anything rather than a code, and have
an in-function check for a code object.

The actual problem arised since coord? is being used a flat-contract
in some places, so it needs to take anything.