Add skeleton g-code-dc% class.
We want to be able to convert picts to G-code. Since picts
are really just convenient wrappers for dc%s, it is more general
to create a class implementing the dc% interface.
Indeed, converting a pict to an SVG makes use of the svg-dc% class.

The new class implements the bulk of the conversion process. For now,
we plan on a simple conversion process: only paths can be converted, and
fill and strokes are completely ignored. Much of the functionality dc%s
support will be unsupported for now.

The end result should be a list of commands as well. This makes it easy
for the user to use other functions we provide before writing.